Fix compilation of verbs provider with ofed 3.12-1
This commit works around the multiple definition errors and allows compilation on RHEL 6.6 without
modifying the installed headers as described in the README.

The correct fix for this is to modify the makefiles to properly depend on the compat-rdma kernel
modules/symbols. I can't figure this out however and the only apparent documentation (OFED_tips.txt
section 4) is out of date and wrong.

Updated the TODO and README to catpure this info.

Signed-off-by: pmmccorm <patrick.m.mccormick@intel.com>